centralize role checks
