+++ apps/web/src/pages/Departments/CreateDepartmentPage.tsx
@@ -14,7 +14,6 @@ import {
   DepartmentSize,
   FragmentType,
 } from "@gc-digital-talent/graphql";
-import { ROLE_NAME } from "@gc-digital-talent/auth";
 import { Heading, Link, CardSeparator, Card } from "@gc-digital-talent/ui";
 
 import SEO from "~/components/SEO/SEO";
@@ -22,12 +21,12 @@ import useRoutes from "~/hooks/useRoutes";
 import useBreadcrumbs from "~/hooks/useBreadcrumbs";
 import pageTitles from "~/messages/pageTitles";
 import Hero from "~/components/Hero";
-import { requireUser } from "~/routing/auth";
 import { graphqlClientContext } from "~/routing/context";
 
 import type { Route } from "./+types/CreateDepartmentPage";
 import FormFields, { DepartmentFormOptions_Fragment } from "./FormFields";
 import { DepartmentType, departmentTypeToInput } from "./utils";
+import { checkPlatformAdmin } from "./roleChecks";
 
 interface FormValues {
   name?: LocalizedStringInput;
@@ -156,10 +155,7 @@ const CreateDepartment_Mutation = graphql(/* GraphQL */ `
 `);
 
 export const clientMiddleware: Route.ClientMiddlewareFunction[] = [
-  async ({ context, request }, next) => {
-    requireUser(context, request, [{ name: ROLE_NAME.PlatformAdmin }]);
-    return await next();
-  },
+  checkPlatformAdmin,
 ];
 
 export async function clientLoader({ context }: Route.ClientLoaderArgs) {
