+++ apps/web/src/pages/Departments/utils.tsx
@@ -150,11 +150,3 @@ export function roleAssignmentsToRoleDepartmentArray(
 
   return collection;
 }
-
-export const DepartmentTeams_Query = graphql(/** GraphQL */ `
-  query DepartmentTeams($id: UUID!) {
-    department(id: $id) {
-      teamIdForRoleAssignment
-    }
-  }
-`);
